requirements.txt
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.0.1
itsdangerous==2.1.2
jaraco.classes==3.3.1
jeepney==0.8.0
Jinja2==3.1.3
keyring==24.3.1
MarkupSafe==2.1.5
more-itertools==10.2.0
paramiko==3.4.0
psycopg2-binary==2.9.9
pycparser==2.21
PyNaCl==1.5.0
SecretStorage==3.3.3
SQLAlchemy==2.0.27
typing_extensions==4.10.0
Werkzeug==3.0.1
zipp==3.17.0
